refactor: move constants, hooks, types folders under lib folder
