+++ app/(tabs)/home/city-trending.tsx
@@ -1,6 +1,6 @@
 import { SafeAreaView, ScrollView, View } from "react-native";
-import { usePlaceListData } from "@/hooks/query/placeData";
-import { PlaceType } from "@/types/api/place";
+import { usePlaceListData } from "@/lib/hooks/query/placeData";
+import { PlaceType } from "@/lib/types/api/place";
 import CityCard from "@/components/ui/home/list/CityCard";
 
 export default function CityTrendingScreen() {
