+++ app/api/reviews/reviews+api.ts
@@ -1,4 +1,4 @@
-import { ReviewSearchParams, ReviewSort } from "@/types/api/review";
+import { ReviewSearchParams, ReviewSort } from "@/lib/types/api/review";
 
 export async function GET(request: Request) {
   const searchParams = new URLSearchParams();
