+++ components/common/input/SearchBar.tsx
@@ -1,8 +1,8 @@
 import { View } from "react-native";
 import { TextInput } from "react-native-gesture-handler";
-import useSearchInput from "@/hooks/common/useSearchInput";
+import useSearchInput from "@/lib/hooks/common/useSearchInput";
 import SearchIcon from "@/assets/icons/common/search.svg";
-import COLOR from "@/constants/colors";
+import COLOR from "@/lib/constants/colors";
 
 export default function SearchBar() {
   const textInputProps = useSearchInput();
