+++ components/ui/home/index/Section.tsx
@@ -1,5 +1,5 @@
 import { View } from "react-native";
-import type { HomeSectionId } from "@/types/screen/home";
+import type { HomeSectionId } from "@/lib/types/screen/home";
 import HomeSectionHeader from "@/components/ui/home/index/molecules/SectionHeader";
 import HomeSectionList from "@/components/ui/home/index/molecules/SectionList";
