+++ components/ui/home/index/molecules/SectionList.tsx
@@ -1,9 +1,9 @@
 import { View } from "react-native";
-import { useSectionData } from "@/hooks/query/sectionData";
-import type { HomeSectionItemProps } from "@/types/screen/home";
-import type { PlaceData } from "@/types/api/place";
-import type { ReviewData } from "@/types/api/review";
-import type { TravelLogData } from "@/types/api/travelLog";
+import { useSectionData } from "@/lib/hooks/query/sectionData";
+import type { HomeSectionItemProps } from "@/lib/types/screen/home";
+import type { PlaceData } from "@/lib/types/api/place";
+import type { ReviewData } from "@/lib/types/api/review";
+import type { TravelLogData } from "@/lib/types/api/travelLog";
 import HorizontalScrollView from "@/components/common/layout/HorizontalScrollView";
 import PlaceCard from "@/components/ui/home/index/atoms/PlaceCard";
 import CityCard from "@/components/ui/home/index/atoms/CityCard";
