+++ components/ui/home/list/RecommendCard.tsx
@@ -1,5 +1,5 @@
 import { Text, View } from "react-native";
-import type { NearByPlaceData } from "@/types/api/nearby";
+import type { NearByPlaceData } from "@/lib/types/api/nearby";
 import StarRating from "@/components/common/display/StarRating";
 import TagChip from "@/components/common/display/TagChip";
