+++ lib/constants/colors.js
