+++ lib/hooks/common/useSearchInput.tsx
