+++ lib/types/screen/travelReview.ts
@@ -1,3 +1,3 @@
-import { ReviewSort } from "@/types/api/review";
+import { ReviewSort } from "@/lib/types/api/review";
 
 export type SortOption = ReviewSort.NEW | ReviewSort.HOT;
